测试移动类
添加补帧移动，测试行动类

+++ Assets/Scripts/Logic/Config/LogicFrameSyncConfig.cs
@@ -14,7 +14,7 @@
         /// <summary>
         /// 逻辑帧间隔Ms 
         /// </summary>
-        public static float LogicFrameIntervalMs = 66f;
+        public static int LogicFrameIntervalMs = 66;
 
 
 }
